refactor: change loop counters from size_t to uint8_t in example models

+++ examples/fourier_transform_1d.cpp
@@ -24,8 +24,8 @@ class TightBindingModel1D {
   Expression m_hamiltonian;
 
   void construct_hamiltonian() {
-    for (size_t i = 0; i < m_size; ++i) {
-      size_t j = (i + 1) % m_size;  // Periodic boundary conditions
+    for (uint8_t i = 0; i < m_size; ++i) {
+    uint8_t j = (i + 1) % m_size;  // Periodic boundary conditions
       m_hamiltonian += Expression::hopping(i, j, Operator::Spin::Up) * m_t;
       m_hamiltonian += Expression::hopping(i, j, Operator::Spin::Down) * m_t;
     }
@@ -44,7 +44,7 @@ static Expression fourier_transform_operator(const Operator& op, size_t size) {
         0.0f, -type_sign * factor * static_cast<float>(k * op.orbital()));
     coefficient = std::exp(coefficient) / std::sqrt(static_cast<float>(size));
 
-    Operator transformed_op(op.type(), op.spin(), k);
+    Operator transformed_op(op.type(), op.spin(), static_cast<uint8_t>(k));
     result += Term(coefficient, {transformed_op});
   }
